% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{arrange}
\alias{arrange}
\title{arrange}
\usage{
arrange(beta, a, manifoldfix = 1)
}
\arguments{
\item{beta}{the tensor in CP components form.}

\item{a}{is the scale.}
}
\value{
the absorbed CP components. (list)
}
\description{
absorb a scale \eqn{a} to the tensor in CP components
}
