% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{ctprod}
\alias{ctprod}
\title{ctprod}
\usage{
ctprod(A, B, K)
}
\arguments{
\item{A}{a tensor}

\item{B}{a tensor}

\item{K}{the inner product \eqn{< >_K}}
}
\value{
contract inner product
}
\description{
contract inner product of two tensor; developed in other package.
}
\references{
Lock, E. F. (2018) Tensor-on-tensor regression. Journal of Computational and Graphical Statis- tics, 27, 638–647. https://github.com/lockEF/MultiwayRegression.
}
