% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{dualascentforalpha}
\alias{dualascentforalpha}
\title{dualascentforalpha}
\usage{
dualascentforalpha(
  y,
  DX,
  alphaold,
  C = 1,
  Maxiterfordual = 1000,
  TolFundualascent = 1e-07,
  MaxIterInner = 40
)
}
\description{
dual ascent method to update the coefficients of basis direction.
}
