% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{elastic_gradient_d}
\alias{elastic_gradient_d}
\title{elastic_gradient_d}
\usage{
elastic_gradient_d(betad, alpha, lambda)
}
\description{
the rescale factor for elastic this is used for Newton method in rescale strategy
}
