% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{elastic_newton_r}
\alias{elastic_newton_r}
\title{elastic_newton_r}
\usage{
elastic_newton_r(beta, alpha, d, rr)
}
\description{
the function used in rescale strategy; Newton's method for r return the lagrange factor
}
