% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{fhatnorm_ten}
\alias{fhatnorm_ten}
\title{fhatnorm_ten}
\usage{
fhatnorm_ten(BB, knots = c(0, 0.25, 0.5, 0.75, 1), order = 4)
}
\arguments{
\item{BB}{the estimated coeffcient tensor in array form.}

\item{knots}{the knots of the spline basis}

\item{order}{the order of the spline basis}
}
\value{
the norm tensor
}
\description{
calculate the norm tensor for BroadcasTR
}
\references{
Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
