% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{fhatsqure_int}
\alias{fhatsqure_int}
\title{fhatsqure_int}
\usage{
fhatsqure_int(coef, knots = c(0, 0.25, 0.5, 0.75, 1), order = 4)
}
\arguments{
\item{coef}{the coefficient of truncated power basis (without constant term)}

\item{knots}{the knots of the spline basis}

\item{order}{the order of the spline basis}
}
\value{
\eqn{\int fhat^2 (entry-wize)}
}
\description{
calculate \eqn{\int fhat^2 (entry-wize)}
}
