% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{full_R}
\alias{full_R}
\title{full_R}
\usage{
full_R(Beta)
}
\arguments{
\item{Beta}{the tensor in CP components form. Here the CP components hold in an list. Each component takes one slot.}
}
\value{
the tensor in array form
}
\description{
transform the tensor of CP decomponets form to the array form
}
