% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{interpoly_int}
\alias{interpoly_int}
\title{interpoly_int}
\usage{
interpoly_int(t1, m1, t2, m2, right_end = 1)
}
\arguments{
\item{t1}{one knot}

\item{m1}{one order}

\item{t2}{one knot}

\item{m2}{one order}

\item{right_end}{the largest knot.}
}
\value{
\eqn{\int (x-t1)^m1 (x-t2)^m2 dx}
}
\description{
calculate \eqn{\int (x-t1)^m1 (x-t2)^m2 dx}
}
