% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{linearpart_penalty_function}
\alias{linearpart_penalty_function}
\title{linearpart_penalty_function}
\usage{
linearpart_penalty_function(y, X, beta, beta0, lambda)
}
\description{
some function used for debug
}
\references{
Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
