% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{linspace}
\alias{linspace}
\title{linspace}
\usage{
linspace(d1, d2, n)
}
\description{
follow the function used in MATLAB toolbox TensorReg
}
\references{
Zhou, H., Li, L. and Zhu, H. (2013) Tensor regression with applications in neuroimaging data analysis. Journal of the American Statistical Association, 108, 540–552. (https://hua-zhou.github.io/TensorReg/)
}
