% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{loglike}
\alias{loglike}
\title{loglike}
\usage{
loglike(y, linearpart, family = "binomial")
}
\arguments{
\item{y}{all the response}

\item{linearpart}{the linear predictor}
}
\value{
the loss
}
\description{
the loss function; for guassian, it is l2 loss
}
\references{
Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
