% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_warmstart.R
\name{single_warmstart}
\alias{single_warmstart}
\title{single_warmstart}
\usage{
single_warmstart(
  X_sample,
  y,
  r = 2,
  restriction = 0,
  knots = 0,
  penalty = c("L1L2"),
  lambda = 0,
  gamma = 0,
  alpha_gamma = 0,
  alpha = 0,
  family = "gaussian",
  Z_sample = NA,
  B0 = NA,
  BB0 = NA,
  Replicates = 5,
  epsilon = 1e-07,
  MaxIter = 1000,
  TolFun = 0.01,
  Memorysave = 0,
  rescale_l2 = 1,
  rescale_all = 0,
  manifold = 0,
  warmstart = 3,
  improvement = 1,
  stoppingrule_penalty = 1,
  QCQP = 1,
  beta0 = NA,
  beta0_vec = NA,
  shrink_factor_scale = 5
)
}
\description{
one times down-size warmstart strategies, which will be used in the sequential warmstart.
}
\references{
Section D of the supplementary for Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
\seealso{
validation_broadcasted_sparsetenreg, broadcasted_sparsetenreg
}
