% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{tildePhiX_trans}
\alias{tildePhiX_trans}
\title{tildePhiX_trans}
\usage{
tildePhiX_trans(X_sample, knots = c(0, 0.25, 0.5, 0.75, 1), order = 4)
}
\arguments{
\item{X_sample}{The training input without basis transformation.}

\item{knots}{the knots of the spline basis}

\item{order}{the order of the spline basis}
}
\value{
\item{BX}{\eqn{\tilde{\phi}(X)} (truncated power basis without the constant term)}
}
\description{
transform the input \eqn{X} to \eqn{\tilde{\phi}(X)} (truncated power basis without the constant term)
}
\references{
Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
\seealso{
validation_broadcasted_sparsetenreg, broadcasted_sparsetenreg
}
