% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_needed.R
\name{truncatS}
\alias{truncatS}
\title{truncatS}
\usage{
truncatS(x, knots, order)
}
\arguments{
\item{x}{the point}

\item{knots}{the knots of the spline basis}

\item{order}{the order of the spline basis}
}
\value{
truncated power basis function value without constant term
}
\description{
generate one truncated Spline; used for debug.
}
\references{
Y. Zhou, R. K. W. Wong and K. He. Broadcasted Nonparametric Tensor Regression
}
